Python/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ML/NLP Libraries
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1
chromadb==0.4.15

# Database & Cache
pymongo==4.5.0
redis==5.0.1
motor==3.3.2

# Background Tasks
celery==5.3.1
redis==5.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
requests==2.31.0
aiohttp==3.9.1
python-magic==0.4.27
email-validator==2.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.0
pre-commit==3.5.0